% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appmetricaR.R
\name{get_appmetrica_data}
\alias{get_appmetrica_data}
\title{Get data from your AppMetrica application}
\usage{
get_appmetrica_data(ids = NULL, date1 = NULL, date2 = NULL,
  metrics = "ym:ge:users", dimensions = "ym:ge:appID",
  filters = NULL, group = NULL, accuracy = "full",
  include_undefined = NULL, lang = NULL, token = NULL)
}
\arguments{
\item{ids}{Comma-separated list of counter numbers.}

\item{date1}{Start date of the report period in the format YYYY-MM-DD. You can also use the values: today, yesterday, ndaysAgo. Default value: 6daysAgo}

\item{date2}{End date of the report period in the format YYYY-MM-DD. You can also use the values: today, yesterday, ndaysAgo. Default value: today}

\item{metrics}{Comma-separated list of metrics. Limit: 20 metrics per request.}

\item{dimensions}{Comma-separated list of dimensions. Limit: 10 dimensions per request.}

\item{filters}{Segmentation filter. Limit: Maximum of 10 unique dimensions and metrics, 20 separate filters, and 10,000 characters per filter string.}

\item{group}{Grouping data by time. Default value: week}

\item{accuracy}{Accuracy of results. Allows you to manage sampling (the number of visits used to calculate the final value). Default value: full}

\item{include_undefined}{Outputs rows that don't have defined dimension values. This only affects the first dimension. Disabled by default.}

\item{lang}{Language.}

\item{token}{An access token from the Yandex.OAuth service}
}
\value{
The requested data will be returned as a data frame.
}
\description{
The Reporting API allows you to get application traffic statistics and other data without using the AppMetrica interface. You can create the desired report structure by specifying metrics and dimensions in the API request.
}
\examples{
\dontrun{
get_appmetrica_data (
ids = "130830,130825",
date1 = "2018-01-01",
date2 = "2018-02-02",
metrics = "ym:ge:users",
dimensions = "ym:ge:mobileDeviceBranding,ym:ge:mobileDeviceModel,ym:ge:appID,ym:ge:date"
)
}
}
