% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appmetricaR.R
\name{get_appmetrica_logs}
\alias{get_appmetrica_logs}
\title{Get non-aggregated data from your AppMetrica application}
\usage{
get_appmetrica_logs(type = "installations", application_id = NULL,
  date_since = "2018-01-01", date_until = "2018-01-02",
  date_dimension = "default",
  fields = "install_datetime,appmetrica_device_id,is_reinstallation",
  token = NULL)
}
\arguments{
\item{type}{API resources}

\item{application_id}{Unique numeric identifier for the application in AppMetrica}

\item{date_since}{Start of the date range in the yyyy-mm-dd hh:mm:ss format}

\item{date_until}{End of the date range in the yyyy-mm-dd hh:mm:ss format}

\item{date_dimension}{The parameter defines what date and time are used as a condition for getting into the data sample. default вЂ” When the event occurred on the device; receive вЂ” When the server received event information.}

\item{fields}{A comma-separated list of fields for the sample}

\item{token}{An access token from the Yandex.OAuth service}
}
\value{
The requested data will be returned as a data frame.
}
\description{
This function allows you to obtain raw data from your AppMetrica application. The AppMetrica Logs API accepts the request and puts it in the queue. If the request is processed successfully, AppMetrica prepares a file for download. In this case, the API returns the HTTP 202 Accepted status. If the request resulted in an error, the API returns the appropriate response status, and the HTTP message body contains the error description.
}
\examples{
\dontrun{
get_appmetrica_logs (type = 'installations',
application_id = 130000,
date_since = '2019-05-01',
date_until = '2019-05-31',
fields = 'install_datetime,publisher_name,tracker_name,appmetrica_device_id,click_url_parameters,is_reinstallation'
)
}
}
